requirements.txt
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.1.0
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0
langchain-openai>=0.0.5
pypdf>=3.15.1
sentence-transformers>=2.2.2
chromadb>=0.4.18
pydantic>=2.0.0
huggingface_hub>=0.19.0
python-dotenv>=1.0.0
streamlit>=1.30.0
colorama>=0.4.6
pandas>=2.0.0 